requirements.txt
requests==2.19.1
system_hotkey==1.0.3
pytesseract==0.2.4
Pillow==7.1.1
PySide2==5.14.2
